project/requirements.txt
Flask
python-dotenv
openai
psycopg2-binary     